6-11

+++ uaa/src/main/java/com/cross/uaa/domain/User.java
@@ -133,7 +133,7 @@ public class User extends AbstractAuditingEntity implements Serializable {
 
     @ApiModelProperty("累计消费次数")
     @Column(name = "order_total_times")
-    private Integer orderTotalTimes;
+    private Long orderTotalTimes;
 
     @ApiModelProperty("累计消费金额")
     @Column(name = "order_total_amount")
@@ -207,11 +207,11 @@ public class User extends AbstractAuditingEntity implements Serializable {
         this.realNameAuthStatus = realNameAuthStatus;
     }
 
-    public Integer getOrderTotalTimes() {
+    public Long getOrderTotalTimes() {
         return orderTotalTimes;
     }
 
-    public void setOrderTotalTimes(Integer orderTotalTimes) {
+    public void setOrderTotalTimes(Long orderTotalTimes) {
         this.orderTotalTimes = orderTotalTimes;
     }
